Refactor(bert): Restructure runtime with wrapper module, update paths, and simplify READMEs
