+++ runtime/bert/inference_mxq.py
@@ -1,7 +1,7 @@
 from argparse import ArgumentParser
 
 import torch
-from bertmxq import BertMXQModel
+from wrapper.bertmxq import BertMXQModel
 from transformers import BertTokenizer
 
 dummy_corpus = [
@@ -24,21 +24,22 @@ if __name__ == "__main__":
     parser.add_argument(
         "--mxq_path",
         type=str,
-        default="../../../compilation/transformers/bert/stsb-bert-tiny-safetensors.mxq",
+        default="../../compilation/bert/mxq/stsb-bert-tiny-safetensors.mxq",
     )
     parser.add_argument(
         "--weight_path",
         type=str,
-        default="../../../compilation/transformers/bert/weight_dict.pth",
+        default="../../compilation/bert/weights/weight_dict.pth",
     )
     args = parser.parse_args()
 
     model = BertMXQModel(args.mxq_path, args.weight_path)
 
+    print("Cosine Similarity (range: -1 to 1, higher = more similar)\n")
     for dummy_pair in dummy_corpus:
         with torch.no_grad():
             s1 = model(**tokenizer(dummy_pair[0], return_tensors="pt"))
             s2 = model(**tokenizer(dummy_pair[1], return_tensors="pt"))
             similarity = torch.nn.functional.cosine_similarity(s1, s2, dim=0)
 
-        print(similarity)
+        print(f"  {similarity.item():.4f}  |  \"{dummy_pair[0]}\" vs \"{dummy_pair[1]}\"")
